dagster_Job/requirements.txt
dagster
dagit
boto3
requests
dagster-postgres
tensorflow
pydantic
keras
pandas
numpy
scikit-learn
wandb
psycopg2-binary
